requirements.txt
python-telegram-bot==20.7
snscrape
certifi
wikipedia-api
fake-useragent
